separate grpc components
