+++ grpc-api/src/main/proto/api.proto
@@ -1,0 +1,33 @@
+syntax = "proto3";
+
+import "google/api/annotations.proto";
+import "google/protobuf/empty.proto";
+
+option java_multiple_files = true;
+option java_package = "com.quorum.tessera.grpc.api";
+
+service Transaction {
+
+    rpc Send (SendRequest) returns (SendResponse) { }
+
+    rpc Receive (ReceiveRequest) returns (ReceiveResponse) { }
+}
+
+message SendRequest {
+    bytes payload = 1;
+    string from = 2;
+    repeated string to = 3;
+}
+
+message SendResponse {
+    string key = 1;
+}
+
+message ReceiveRequest {
+    string key = 1;
+    string to = 2;
+}
+
+message ReceiveResponse {
+    bytes payload = 1;
+}
