+++ grpc-service/src/main/java/com/quorum/tessera/client/GrpcP2pClient.java
@@ -1,6 +1,6 @@
 package com.quorum.tessera.client;
 
-import com.quorum.tessera.api.grpc.Convertor;
+import com.quorum.tessera.grpc.p2p.Convertor;
 import com.quorum.tessera.api.model.ResendRequest;
 
 import java.util.Objects;
@@ -30,7 +30,7 @@ class GrpcP2pClient implements P2pClient {
 
     @Override
     public boolean makeResendRequest(String targetUrl, ResendRequest request) {
-        com.quorum.tessera.api.grpc.model.ResendRequest grpcObj =  Convertor.toGrpc(request);
+        com.quorum.tessera.grpc.p2p.ResendRequest grpcObj =  Convertor.toGrpc(request);
         return grpcClientFactory.getClient(targetUrl).makeResendRequest(grpcObj);
     }
