+++ grpc-service/src/main/java/com/quorum/tessera/grpc/StreamObserverCallback.java
@@ -1,5 +1,5 @@
 
-package com.quorum.tessera.api.grpc;
+package com.quorum.tessera.grpc;
 
 
 @FunctionalInterface
