+++ grpc-service/src/main/java/com/quorum/tessera/grpc/p2p/PartyInfoGrpcService.java
@@ -1,7 +1,8 @@
-package com.quorum.tessera.api.grpc;
+package com.quorum.tessera.grpc.p2p;
 
 import com.google.protobuf.ByteString;
-import com.quorum.tessera.api.grpc.model.PartyInfoMessage;
+import com.quorum.tessera.grpc.p2p.PartyInfoMessage;
+import com.quorum.tessera.grpc.StreamObserverTemplate;
 import com.quorum.tessera.node.PartyInfoParser;
 import com.quorum.tessera.node.PartyInfoService;
 import com.quorum.tessera.node.model.PartyInfo;
