+++ grpc-service/src/test/java/com/quorum/tessera/grpc/p2p/TransactionGrpcServiceTest.java
@@ -1,8 +1,7 @@
-package com.quorum.tessera.api.grpc;
+package com.quorum.tessera.grpc.p2p;
 
 import com.google.protobuf.ByteString;
 import com.quorum.tessera.transaction.TransactionManagerImpl;
-import com.quorum.tessera.api.grpc.model.*;
 import io.grpc.stub.StreamObserver;
 import java.util.Base64;
 import java.util.Optional;
@@ -18,12 +17,6 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 
 public class TransactionGrpcServiceTest {
-
-    @Mock
-    private StreamObserver<SendResponse> sendResponseObserver;
-
-    @Mock
-    private StreamObserver<ReceiveResponse> receiveResponseObserver;
 
     @Mock
     private StreamObserver<DeleteRequest> deleteResponseObserver;
@@ -48,90 +41,9 @@ public class TransactionGrpcServiceTest {
     @After
     public void onTearDown() {
         verifyNoMoreInteractions(
-                sendResponseObserver,
-                receiveResponseObserver,
                 deleteResponseObserver,
                 pushResponseObserver,
                 resendResponseObserver);
-    }
-
-    @Test
-    public void testSend() {
-
-        SendRequest sendRequest = SendRequest.newBuilder()
-                .setFrom("bXlwdWJsaWNrZXk=")
-                .addTo("cmVjaXBpZW50MQ==")
-                .setPayload(ByteString.copyFromUtf8("Zm9v")).build();
-
-        com.quorum.tessera.api.model.SendResponse r = new com.quorum.tessera.api.model.SendResponse("KEY");
-        when(enclaveMediator.send(any())).thenReturn(r);
-
-        service.send(sendRequest, sendResponseObserver);
-
-        verify(enclaveMediator).send(any());
-
-        verify(sendResponseObserver).onNext(any());
-
-        verify(sendResponseObserver).onCompleted();
-    }
-
-    @Test
-    public void testSendWithEmptySender() {
-        SendRequest sendRequest = SendRequest.newBuilder()
-                .addTo("cmVjaXBpZW50MQ==")
-                .setPayload(ByteString.copyFromUtf8("Zm9v"))
-                .build();
-
-        com.quorum.tessera.api.model.SendResponse r = new com.quorum.tessera.api.model.SendResponse("KEY");
-        when(enclaveMediator.send(any())).thenReturn(r);
-
-        service.send(sendRequest, sendResponseObserver);
-
-        verify(enclaveMediator).send(any());
-
-        verify(sendResponseObserver).onNext(any());
-
-        verify(sendResponseObserver).onCompleted();
-    }
-
-    @Test
-    public void testReceive() {
-
-        com.quorum.tessera.api.model.ReceiveResponse r = new com.quorum.tessera.api.model.ReceiveResponse("SOME DATA".getBytes());
-
-        
-        when(enclaveMediator.receive(any())).thenReturn(r);
-
-        ReceiveRequest request = ReceiveRequest.newBuilder()
-                .setTo("cmVjaXBpZW50MQ==")
-                .setKey("ROAZBWtSacxXQrOe3FGAqJDyJjFePR5ce4TSIzmJ0Bc=")
-                .build();
-
-        service.receive(request, receiveResponseObserver);
-
-        verify(enclaveMediator).receive(any());
-
-        ArgumentCaptor<ReceiveResponse> receiveResponseCaptor = ArgumentCaptor.forClass(ReceiveResponse.class);
-        verify(receiveResponseObserver).onNext(receiveResponseCaptor.capture());
-        ReceiveResponse response = receiveResponseCaptor.getValue();
-
-        assertThat(response).isNotNull();
-        assertThat(response.getPayload().toStringUtf8()).isEqualTo("SOME DATA");
-        verify(receiveResponseObserver).onCompleted();
-    }
-
-    @Test
-    public void testReceiveWithNoToField() {
-
-        ReceiveRequest request = ReceiveRequest.newBuilder()
-                .setKey("ROAZBWtSacxXQrOe3FGAqJDyJjFePR5ce4TSIzmJ0Bc=")
-                .build();
-
-
-        service.receive(request, receiveResponseObserver);
-        
-
-        verify(receiveResponseObserver).onError(any());
     }
 
     @Test
@@ -192,32 +104,6 @@ public class TransactionGrpcServiceTest {
 
         verify(resendResponseObserver).onNext(any());
         verify(resendResponseObserver).onCompleted();
-    }
-
-
-    @Test
-    public void invalidSendRequest() {
-
-        SendRequest sendRequest = SendRequest.newBuilder()
-                .setFrom("bXlwdWJsaWNrZXk=")
-                .addTo("cmVjaXBpZW50MQ==").build();
-
-        service.send(sendRequest, sendResponseObserver);
-
-        verify(sendResponseObserver).onError(any());
-
-    }
-
-    @Test
-    public void invalidReceiveRequest() {
-
-        ReceiveRequest request = ReceiveRequest.newBuilder()
-                .build();
-
-        service.receive(request, receiveResponseObserver);
-
-        verify(receiveResponseObserver).onError(any());
-
     }
 
     @Test
