+++ tessera-core/src/test/resources/H2-byteFunctions.sql
@@ -1,0 +1,4 @@
+CREATE ALIAS bytesToHex FOR "javax.xml.bind.DatatypeConverter.printHexBinary";
+CREATE ALIAS bytesToB64 FOR "javax.xml.bind.DatatypeConverter.printBase64Binary";
+CREATE ALIAS hexToBytes FOR "javax.xml.bind.DatatypeConverter.parseHexBinary";
+CREATE ALIAS b64ToBytes FOR "javax.xml.bind.DatatypeConverter.parseBase64Binary";
